Error handling and stability Improvements
Now even if some part of test fails It'll keep going but throw a fail and create a file with the failures and where and how they happened. The failOnPurpose.js file is for testing to see if this system works and will cause the program to throw a fail after all scripts run and won't be a part of the final script. Please give me feedback on what to add/update next.